Move configuration to models and add sso related values
