+++ Models/Configuration_test.go
@@ -1,4 +1,4 @@
-package main
+package Models
 
 import (
 	"github.com/sirupsen/logrus"
@@ -17,6 +17,7 @@ func TestNewConfiguration_AllDefaults(t *testing.T) {
 	assert.Equal(t, logrus.InfoLevel, config.LogLevel)
 	assert.Nil(t, config.TrustedProxies)
 	assert.Equal(t, "localhost:5000", config.ListenAddress)
+	assert.Equal(t, "rinkudesu", config.SsoClientId)
 }
 
 func TestNewConfiguration_CustomValues(t *testing.T) {
@@ -26,6 +27,8 @@ func TestNewConfiguration_CustomValues(t *testing.T) {
 	_ = os.Setenv("TAGS_DB", "postgres://postgres:postgres@localhost:5432/postgres")
 	_ = os.Setenv("TAGS_PROXY", "192.168.0.1,10.0.0.1,10.0.0.2")
 	_ = os.Setenv("TAGS_ADDRESS", "192.168.0.1:80")
+	_ = os.Setenv("TAGS_AUTHORITY", "http://localhost")
+	_ = os.Setenv("TAGS_CLIENTID", "not-rinkudesu")
 
 	config := NewConfiguration()
 
@@ -34,4 +37,6 @@ func TestNewConfiguration_CustomValues(t *testing.T) {
 	assert.Equal(t, "postgres://postgres:postgres@localhost:5432/postgres", config.DbConnection)
 	assert.Equal(t, []string{"192.168.0.1", "10.0.0.1", "10.0.0.2"}, config.TrustedProxies)
 	assert.Equal(t, "192.168.0.1:80", config.ListenAddress)
+	assert.Equal(t, "http://localhost", config.SsoAuthority)
+	assert.Equal(t, "not-rinkudesu", config.SsoClientId)
 }
